WIP: Use smallest numeric literal at the sql parser
This change will use the smallest numeric literal when converting untyped literal
inputs at the sql parser. Till now a numeric literal input resulted either in a
`LongLiteral` or `DoubleLiteral`.
A special logic inside the (old) function parameter binding tried to downcast
these literals later on which resulted in bugs on nested function expressions.
See https://github.com/crate/crate/issues/9652.
Using smallest possible numeric literal at the parser will supersede such logic
while ensuring correct function matching also on nested expressions.

+++ sql/src/main/java/io/crate/analyze/NegateLiterals.java
@@ -56,7 +56,7 @@ public final class NegateLiterals extends SymbolVisitor<Void, Symbol> {
             case FloatType.ID:
                 return Literal.of(valueType, (Double) value * -1);
             case ShortType.ID:
-                return Literal.of(valueType, (Short) value * -1);
+                return Literal.of(valueType, ((Integer) ((Short) value * -1)).shortValue());
             case IntegerType.ID:
                 return Literal.of(valueType, (Integer) value * -1);
             case LongType.ID:
